requirements.txt
feedparser==6.0.10
python-docx==0.8.11
PyPDF2==3.0.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
torch==2.0.1
pytest==7.4.0
tqdm==4.66.1 